requirements.txt
Jinja2==3.1.2
jsonschema==4.10.2
pyaml==21.10.1
PyYAML==6.0
ansible